Add support for multiple cpi versions
[156356323](https://www.pivotaltracker.com/story/show/156356323)
